+++ src/pipeline.rs
@@ -1,7 +1,6 @@
 use std::borrow::Cow;
 
-use bevy::ecs::query::QueryItem;
-use bevy::ecs::system::lifetimeless::Read;
+use bevy::ecs::system::lifetimeless::SQuery;
 use bevy::ecs::system::SystemParamItem;
 use bevy::pbr::{
     setup_morph_and_skinning_defs, MeshFlags, MeshPipelineKey, MeshPipelineViewLayoutKey,
@@ -10,11 +9,10 @@ use bevy::pbr::{
 use bevy::prelude::*;
 use bevy::render::batching::GetBatchData;
 use bevy::render::render_resource::{
-    BindGroupLayout, BindGroupLayoutDescriptor, BindGroupLayoutEntry, BindingType, BlendState,
-    BufferBindingType, BufferSize, ColorTargetState, ColorWrites, CompareFunction, DepthBiasState,
-    DepthStencilState, Face, FragmentState, FrontFace, MultisampleState, PolygonMode,
-    PrimitiveState, PrimitiveTopology, ShaderDefVal, ShaderSize, ShaderStages, StencilState,
-    TextureFormat, VertexState,
+    BindGroupLayout, BindGroupLayoutEntry, BindingType, BlendState, BufferBindingType, BufferSize,
+    ColorTargetState, ColorWrites, CompareFunction, DepthBiasState, DepthStencilState, Face,
+    FragmentState, FrontFace, MultisampleState, PolygonMode, PrimitiveState, PrimitiveTopology,
+    ShaderDefVal, ShaderSize, ShaderStages, ShaderType, StencilState, TextureFormat, VertexState,
 };
 use bevy::render::renderer::RenderDevice;
 use bevy::render::settings::WgpuSettings;
@@ -67,6 +65,7 @@ impl PipelineKey {
         pub offset_zero, set_offset_zero: 13;
         pub hdr_format, set_hdr_format: 14;
         pub morph_targets, set_morph_targets: 15;
+        view_key_int, set_view_key_int: 23, 16;
     }
 
     pub(crate) fn new() -> Self {
@@ -145,6 +144,15 @@ impl PipelineKey {
         self.set_morph_targets(morph_targets);
         self
     }
+
+    pub(crate) fn with_view_key(mut self, view_key: MeshPipelineViewLayoutKey) -> Self {
+        self.set_view_key_int(view_key.bits());
+        self
+    }
+
+    pub(crate) fn view_key(&self) -> MeshPipelineViewLayoutKey {
+        MeshPipelineViewLayoutKey::from_bits(self.view_key_int()).unwrap()
+    }
 }
 
 impl From<PipelineKey> for MeshPipelineKey {
@@ -171,78 +179,72 @@ impl FromWorld for OutlinePipeline {
         let world = world.cell();
         let mesh_pipeline = world.get_resource::<MeshPipeline>().unwrap().clone();
         let render_device = world.get_resource::<RenderDevice>().unwrap();
-        let outline_view_bind_group_layout =
-            render_device.create_bind_group_layout(&BindGroupLayoutDescriptor {
-                label: Some("outline_view_bind_group_layout"),
-                entries: &[BindGroupLayoutEntry {
+        let outline_view_bind_group_layout = render_device.create_bind_group_layout(
+            "outline_view_bind_group_layout",
+            &[BindGroupLayoutEntry {
+                binding: 0,
+                visibility: ShaderStages::VERTEX,
+                ty: BindingType::Buffer {
+                    ty: BufferBindingType::Uniform,
+                    has_dynamic_offset: true,
+                    min_binding_size: Some(OutlineViewUniform::min_size()),
+                },
+                count: None,
+            }],
+        );
+        let outline_volume_bind_group_layout = render_device.create_bind_group_layout(
+            "outline_volume_bind_group_layout",
+            &[
+                BindGroupLayoutEntry {
                     binding: 0,
                     visibility: ShaderStages::VERTEX,
                     ty: BindingType::Buffer {
                         ty: BufferBindingType::Uniform,
                         has_dynamic_offset: true,
-                        min_binding_size: BufferSize::new(OutlineViewUniform::SHADER_SIZE.get()),
+                        min_binding_size: BufferSize::new(OutlineVolumeUniform::SHADER_SIZE.get()),
                     },
                     count: None,
-                }],
-            });
-        let outline_volume_bind_group_layout =
-            render_device.create_bind_group_layout(&BindGroupLayoutDescriptor {
-                label: Some("outline_volume_bind_group_layout"),
-                entries: &[
-                    BindGroupLayoutEntry {
-                        binding: 0,
-                        visibility: ShaderStages::VERTEX,
-                        ty: BindingType::Buffer {
-                            ty: BufferBindingType::Uniform,
-                            has_dynamic_offset: true,
-                            min_binding_size: BufferSize::new(
-                                OutlineVolumeUniform::SHADER_SIZE.get(),
-                            ),
-                        },
-                        count: None,
-                    },
-                    BindGroupLayoutEntry {
-                        binding: 1,
-                        visibility: ShaderStages::FRAGMENT,
-                        ty: BindingType::Buffer {
-                            ty: BufferBindingType::Uniform,
-                            has_dynamic_offset: true,
-                            min_binding_size: BufferSize::new(
-                                OutlineFragmentUniform::SHADER_SIZE.get(),
-                            ),
-                        },
-                        count: None,
-                    },
-                ],
-            });
-        let outline_stencil_bind_group_layout =
-            render_device.create_bind_group_layout(&BindGroupLayoutDescriptor {
-                label: Some("outline_stencil_bind_group_layout"),
-                entries: &[BindGroupLayoutEntry {
-                    binding: 0,
-                    visibility: ShaderStages::VERTEX,
+                },
+                BindGroupLayoutEntry {
+                    binding: 1,
+                    visibility: ShaderStages::FRAGMENT,
                     ty: BindingType::Buffer {
                         ty: BufferBindingType::Uniform,
                         has_dynamic_offset: true,
-                        min_binding_size: BufferSize::new(OutlineStencilUniform::SHADER_SIZE.get()),
+                        min_binding_size: BufferSize::new(
+                            OutlineFragmentUniform::SHADER_SIZE.get(),
+                        ),
                     },
                     count: None,
-                }],
-            });
-        let outline_deform_bind_group_layout =
-            render_device.create_bind_group_layout(&BindGroupLayoutDescriptor {
-                label: Some("outline_deform_bind_group_layout"),
-                entries: &[BindGroupLayoutEntry {
-                    binding: 0,
-                    visibility: ShaderStages::VERTEX,
-                    ty: BindingType::Buffer {
-                        ty: BufferBindingType::Uniform,
-                        has_dynamic_offset: true,
-                        min_binding_size: BufferSize::new(OutlineDeformUniform::SHADER_SIZE.get()),
-                    },
-                    count: None,
-                }],
-            });
+                },
+            ],
+        );
+        let outline_stencil_bind_group_layout = render_device.create_bind_group_layout(
+            "outline_stencil_bind_group_layout",
+            &[BindGroupLayoutEntry {
+                binding: 0,
+                visibility: ShaderStages::VERTEX,
+                ty: BindingType::Buffer {
+                    ty: BufferBindingType::Uniform,
+                    has_dynamic_offset: true,
+                    min_binding_size: BufferSize::new(OutlineStencilUniform::SHADER_SIZE.get()),
+                },
+                count: None,
+            }],
+        );
+        let outline_deform_bind_group_layout = render_device.create_bind_group_layout(
+            "outline_deform_bind_group_layout",
+            &[BindGroupLayoutEntry {
+                binding: 0,
+                visibility: ShaderStages::VERTEX,
+                ty: BindingType::Buffer {
+                    ty: BufferBindingType::Uniform,
+                    has_dynamic_offset: true,
+                    min_binding_size: BufferSize::new(OutlineDeformUniform::SHADER_SIZE.get()),
+                },
+                count: None,
+            }],
+        );
         OutlinePipeline {
             mesh_pipeline,
             outline_view_bind_group_layout,
@@ -271,25 +273,19 @@ impl SpecializedMeshPipeline for OutlinePipeline {
             buffer_attrs.push(Mesh::ATTRIBUTE_POSITION.at_shader_location(0));
         }
 
-        let mut bind_layouts = vec![self
-            .mesh_pipeline
-            .get_view_layout(if key.msaa() == Msaa::Off {
-                MeshPipelineViewLayoutKey::empty()
-            } else {
-                MeshPipelineViewLayoutKey::MULTISAMPLED
-            })
-            .clone()];
+        let mut bind_layouts = vec![
+            self.mesh_pipeline.get_view_layout(key.view_key()).clone(),
+            setup_morph_and_skinning_defs(
+                &self.mesh_pipeline.mesh_layouts,
+                layout,
+                5,
+                &key.into(),
+                &mut vertex_defs,
+                &mut buffer_attrs,
+            ),
+            self.outline_view_bind_group_layout.clone(),
+        ];
 
-        bind_layouts.push(setup_morph_and_skinning_defs(
-            &self.mesh_pipeline.mesh_layouts,
-            layout,
-            5,
-            &key.into(),
-            &mut vertex_defs,
-            &mut buffer_attrs,
-        ));
-
-        bind_layouts.push(self.outline_view_bind_group_layout.clone());
         let cull_mode;
         if key.depth_mode() == DepthMode::Flat {
             let val = ShaderDefVal::from("FLAT_DEPTH");
@@ -388,21 +384,20 @@ impl SpecializedMeshPipeline for OutlinePipeline {
 }
 
 impl GetBatchData for OutlinePipeline {
-    type Param = ();
-    type Query = Read<ExtractedOutline>;
-    type QueryFilter = ();
+    type Param = SQuery<&'static ExtractedOutline>;
     type CompareData = ();
     type BufferData = MeshUniform;
 
     fn get_batch_data(
-        _: &SystemParamItem<Self::Param>,
-        outline: &QueryItem<Self::Query>,
-    ) -> (Self::BufferData, Option<Self::CompareData>) {
+        param: &SystemParamItem<Self::Param>,
+        entity: Entity,
+    ) -> Option<(Self::BufferData, Option<Self::CompareData>)> {
+        let outline = param.get(entity).unwrap();
         let ts = MeshTransforms {
             transform: (&outline.transform).into(),
             previous_transform: (&outline.transform).into(),
             flags: MeshFlags::NONE.bits(),
         };
-        ((&ts).into(), None)
+        Some((MeshUniform::new(&ts, None), None))
     }
 }